scheduing considering node utilization

+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-server/hadoop-yarn-server-resourcemanager/src/main/java/org/apache/hadoop/yarn/server/resourcemanager/scheduler/fair/FairSchedulerConfiguration.java
@@ -106,11 +106,11 @@ public class FairSchedulerConfiguration extends Configuration {
   /** Scheduler threshold for node utils. */
   protected static final String LOAD_SCHEDULING_THRESHOLD_NODE_CPU_UTIL =
       CONF_PREFIX + "load.scheduling.threshold.node.cpu.util";
-  protected static final float DEFAULT_LOAD_SCHEDULING_THRESHOLD_NODE_CPU_UTIL = 0.95;
+  protected static final double DEFAULT_LOAD_SCHEDULING_THRESHOLD_NODE_CPU_UTIL = 0.95;
 
   protected static final String LOAD_SCHEDULING_THRESHOLD_NODE_MEMORY_UTIL =
       CONF_PREFIX + "load.scheduling.threshold.node.memory.util";
-  protected static final float DEFAULT_LOAD_SCHEDULING_THRESHOLD_NODE_MEMORY_UTIL = 0.95;
+  protected static final double DEFAULT_LOAD_SCHEDULING_THRESHOLD_NODE_MEMORY_UTIL = 0.95;
 
   /** Cluster threshold for rack locality. */
   protected static final String LOCALITY_THRESHOLD_RACK = CONF_PREFIX + "locality.threshold.rack";
@@ -330,13 +330,13 @@ public class FairSchedulerConfiguration extends Configuration {
     return getFloat(LOCALITY_THRESHOLD_NODE, DEFAULT_LOCALITY_THRESHOLD_NODE);
   }
 
-  public float getLoadSchedulingThresholdNodeCpuUtil() {
-    return getFloat(LOAD_SCHEDULING_THRESHOLD_NODE_CPU_UTIL,
-        DEFAULT_LOAD_SCHEDULING_THRESHOLD_NODE_CPU_UTIL;
+  public double getLoadSchedulingThresholdNodeCpuUtil() {
+    return getDouble(LOAD_SCHEDULING_THRESHOLD_NODE_CPU_UTIL,
+        DEFAULT_LOAD_SCHEDULING_THRESHOLD_NODE_CPU_UTIL);
   }
 
-  public float getLoadSchedulingThresholdNodeMemoryUtil() {
-    return getFloat(LOAD_SCHEDULING_THRESHOLD_NODE_MEMORY_UTIL,
+  public double getLoadSchedulingThresholdNodeMemoryUtil() {
+    return getDouble(LOAD_SCHEDULING_THRESHOLD_NODE_MEMORY_UTIL,
         DEFAULT_LOAD_SCHEDULING_THRESHOLD_NODE_MEMORY_UTIL);
   }
